requirements.txt
streamlit
torch
torchvision
pillow
requests
python-dotenv
groq
